change select control behavior, don't change on keyup event, only change event, also force blur() after a change
before, the focus was left on the select element and changes from other systems would not change
the focused select, then when focus left the select it would force another incorrect change to the sbelement
now, the select is unfocused after it is changed, which results in correct behavior,
except in the case where the select is opened and then closed without changing the option,
it will retain focus then, but when unfocused it will not force a sbelement change, so this is probably
as good as it will get for the select element.  Once jQueryUI releases their Menu widget, all select
elements should be changed to the Menu widgets.