Time Update Throttling
Update ScoreBoardXmlListener to attach new Processing Instruction
"TimeUpdate" with value either "sec", if time changes second
value or if time is at clock start or end, or "ms" if time is
only a change in milliseconds

Update XmlScoreBoardServlet to optionally provide XPath filter to
associated listener that filters out any time updates with
"TimeUpdate" Processing Instructions with value "ms".  Default
is to filter ms updates out; this can be changed at the URL
/XmlScoreBoard/config?timeUpdate=ms when also including the
current registration id as a param.