fix issue #4 by deferring team table creation
the team page has a long freeze when loading because every team's roster editor is generated when the page first loads, and this html is then hideen/shown by the dropdown.  Additionally, each team table takes about 500-600ms to generate, as there are a large number of expensive-to-generate $sbControls in each team's roster editor.  By caching and deferring the team page creation until the team is selected from the dropdown, this update trades off the huge up-front generation cost with smaller freezes when a team is first selected.

Speeding up the roster generation procedure would be preferred, but that requires basically re-writing $sbControl which is a bit too scary for me at this point.