Update README with revised content from wiki.
The content in the wiki has undergone several revisions since 3.9.3 was released. Some changes have been made to the Installing the Scoreboard section and the Quick Start Guide to Operating the Scoreboard has been revised to simplify the language and make terminology more consistent. The Installing the Scoreboard section also includes a note on compiling the .jar from domagalski (issue #40).