Fix Box button not working after Star Pass
When the skater changes positiosn from "Pivot" to "Jammer", the state is not correctly sync-ed between the PositionModel and the SkaterModel.  At Jam end, the Jammer position's PenaltyBox status is cleared, but then not re-set based on the SkaterModel.

This is now "fixed", but this PositionModel/SkaterModel nonsense is a mess, and in an ideal world I would just maintain the state on the SkaterModel and fire Position updates if the skater's status changes.  I would be more comfortable making those changes once unit tests are in place though (#72).

Closes #5