add "unit" tests for DefaultPositionModel
they aren't very unit-y, they depend rather heavily on the Team and Skater Models to work correctly, but I think this accomplishes good coverage of the Position interface.  Added Mockito to avoid pulling in too much of the universe to do these tests.  Near term plan (sometime in the few weeks) is to cook up a nice fixture to test the event system, add Tests for Skater and Team models, then start teasing apart the interweaving of the Team, Skater and Position models.