Fix Box button not working after Star Pass (#78)
* Fix Box button not working after Star Pass

When the skater changes positiosn from "Pivot" to "Jammer", the state is not correctly sync-ed between the PositionModel and the SkaterModel.  At Jam end, the Jammer position's PenaltyBox status is cleared, but then not re-set based on the SkaterModel.

This is now "fixed", but this PositionModel/SkaterModel nonsense is a mess, and in an ideal world I would just maintain the state on the SkaterModel and fire Position updates if the skater's status changes.  I would be more comfortable making those changes once unit tests are in place though (#72).

Closes #5

* add "unit" tests for DefaultPositionModel

they aren't very unit-y, they depend rather heavily on the Team and Skater Models to work correctly, but I think this accomplishes good coverage of the Position interface.  Added Mockito to avoid pulling in too much of the universe to do these tests.  Near term plan (sometime in the few weeks) is to cook up a nice fixture to test the event system, add Tests for Skater and Team models, then start teasing apart the interweaving of the Team, Skater and Position models.

* fix some issues with ant build with new tests