Fix issues when Jam Clock is counting Up (#90)
* add unit tests for DefaultClockModel

doesn't cover a lot of cases related to start()/stop() and the underlying
UpdateClockTimerTask.  That is a task for another day/PR, unfortunately.

* fix warnings in DefaultClockModel

* fix warnings in DefaultScoreBoardModel

* fix bugs with clock ending on clock counting up/down

replace a bunch of getMax/getMin instances with isTimeAtStart/isTimeAtEnd calls.  The "business rules" around clocks is a bit of a mudball at the moment, and kind of hard to test unfortunately.

* fix from code review