completely remove the media-src folder
a media release hasn't been put together in years, so we don't need to retain these items. Beyond the team logos, the only other item in this folder is a canadian flag image/video that again, we haven't made a separate release for since 2012 or so.  Also removed references to the media-src directory from build.xml