Use snapshots for undo
Whenever one of the clock control buttons is pressed (Start Jam, Stop
Jam/TO, Timeout), a snapshot of the scoreboard state is taken. On undo
the last snapshot is restored and the running clocks are advanced by the
time that has passed since the snapshot was taken.
Un-Start, Un-Stop and Un-Timeout are wrappers for the general undo
function that check if the current snapshot is of the approriate type
and do nothing if it isn't.

This prepares the backend for a single undo button but doesn't add it to
the frontend. The latter is planned as an option, after we have a
better way of handling operator settings than a long row of buttons.