Factor out WS change handling into a new class.
This is a better seperation of concerns, that'll allow
for easier unittesting and the json to be reused in other places.
It's more efficient, as the calculation of what has changes is
only done once.

This breaks WS custom settings, used by some overlays.
This removes the WS register "path" argument, which was unused.
This removes stateId from the WS, which was unused and is no longer
needed internally.