Simplify ScoreBoardManager
Hardcode the controllers and viewers to use, which makes
the code easier to understand and allows future development.
This is one more set of constants removed from the properties file.

I would have gone further, however there's a layering violation with
TwitterXmlDocumentManager that'd need fixing first plus how the servlets
are currently loaded.