Switch to WS-based JSON file output from old game output. (#146)
* Pass skater not the team on removal.

This fixes an exception in the xml/json listener.

* Factor out WS change handling into a new class.

This is a better seperation of concerns, that'll allow
for easier unittesting and the json to be reused in other places.
It's more efficient, as the calculation of what has changes is
only done once.

This breaks WS custom settings, used by some overlays.
This removes the WS register "path" argument, which was unused.
This removes stateId from the WS, which was unused and is no longer
needed internally.

* Implement custom settings in the usual fashion.

These are now auto-saved, and accessible to XML screens.
We may wish to move pure-frontend settings (e.g. should
teams be swapped) to this.

Fixes #12

* Simplify ScoreBoardManager

Hardcode the controllers and viewers to use, which makes
the code easier to understand and allows future development.
This is one more set of constants removed from the properties file.

I would have gone further, however there's a layering violation with
TwitterXmlDocumentManager that'd need fixing first plus how the servlets
are currently loaded.

* Move JSONStateManager management to ScoreBoardManager.

This will allow it to be reused elsewhere, without having
to duplicate the work of managing updates.

* Add JSONStateSnapshotter

This will write out game state to disk similarly
the the exsting Game classes, however:
- In addition to at end of Jam, it's also triggered on
  official score, so any points add after the last jam
  aren't missed.
- It's always turned on, and doesn't depend on Start New Game.
- It doesn't lose state across scoreboard restarts.
- It's using the WS state, with identical output so
  it's not yet another data format to maintain
- The filenames include the period number, jam number,
  and indicate if the score is official. This makes it
  easier to find the file you want.
- The filename includes seconds, so if more than one
  jam ends in a given minute you'll now get data for both.
- It's in the game-data directory rather than GameData,
  to help distinguish the two formats.

An example file name is:

2018-08-25 16_26_03 - Black vs White - Period 1 Jam 12.json

* Remove game snapshots.

* Remove the rest of Game.

The game name is no longer used, so can be removed from the UI.

* Fix regression from #130.

The XML js code really doesn't like being asked to delete
XML nodes it doesn't know about.

Also include team id in WS skater stats.

* Reset final score bool when starting a new game

* Don't reset custom settings on reset. Defaults for interactive overlay.

Custom Settings are for things like broadcast overlay options,
which shouldn't have to be setup again on every game.

Enable the basics for the interactive overlay, as it's probably what
users want and avoids showing a confusing blank screen by default.

* Add unittests for ScoreBoardJSONListener

Fix bugs arising from adding tests

* Use walltime of start of the first jam for json file name.

Remove jam/period/final. This gives one file name per game.
If the time is missing, fallback to current time - this
shouldn't happen in a real game.

* Integrate code review comments.

* Rename CustomSettings to FrontendSettings