Use a single lock for all core classes.
In order to avoid potential for deadlocks, cover all the core logic in
Default*Model with a single lock. This could make reaction times a bit
slower, if multiple clients send state updates at the same time but as
long as all updates are handled quickly, this should be negligible. (We
should, however, not start to compute complex stats under the lock.)

The only remaining calls from code under the lock to code (potentially)
under different locks are to the baseLock in Ruleset.java, which in turn
does not reach other locks, and through the event interface. Since all
event listeners not covered by the coreLock are async and thus the
actual handling of the event will not be done while holding the
coreLock, we can conclude that this change and #149 together do not
introduce any potential deadlocks.