Use git tag as version, and include build info.
This avoids having to remember to update a file on every release, and
when ad-hoc builds are made not at a tag they'll include the most recent
tag, how many commits since then, and the commit hash.

Also always include the build username, host, and time so we
can figure out where a build came from if needed.