Switch operator jammer/pivot to WS, adjust logic.
With AdvanceFielding, the existing code no longer really
works. This is now better, however weird stuff can happen
e.g. you set the pivot and jammer to the same skater
while the fielding hasn't been advanced.

In practice unlikely to be an issue, enough to unbreak
non-ePLT setups and will also work in ePLT setups where the
ePLT is the person entering skaters.