Move Twitter over to WS.
This uses async interfaces, so that talking to Twitter
won't block the scoreboard.

As we're now storing secrets in the state, don't let
them leak to game-data as that's intended to be shared.
They're still in the main autosave though.

Keep login across restarts.
Remove "beta", it's not been for a long time.
Update condition/tweet preview in real time.