Switch save/load to JSON. (#314)
Trim down options to what actually works. There's
no distinction between everything and scoreboard now,
and Pages hasn't existed in a while.

Users should use Start New Game, not resetting manually,
so the reset can be removed.

The full reset doesn't work, as it depends on the scoreboard correctly
resetting itself - which it doesn't completely and in the event of a bug
isn't likely to help you anyway if you've a bug like that (or tried to
merge/load in something that you shouldn't have). Blowing away the save
and restarting will always work though.