Don't show jammers when clocks aren't running.
This was causing jammers to appear when no clocks
were running, e.g. just before a period start when
the intermission clock wasn't running.

That isn't a problem in itself, however if you were coming
into P2 and had lead in the last jam of P1, then lead
would be incorrectly indicated before the jam start.

So go with consistency when we're in lineup,
and don't show the jammer name.