Avoid reverse DNS resolution when looking for URLs.
At most events the scoreboard will not be connected to the internet,
so users end up having to wait for a DNS timeout before the URLs show
at startup or on the index page.

This loses us hostnames, but it's very unlikely that there was a DNS
setup that would have allowed them to work.

We'll still do forward resolution to bind to if hostname is provided on
the command line, but that's going to be rare and also likey something
present in /etc/hosts so it won't hit the network.

Also cleanup and modernise the code generally, and don't exclude v6
addresses.