Fix rounding of clocks
The WS display would use the same rounding function as the XML display
but without the $sb objects, causing it to always round up instead of
opposite to the clock direction. Change it to use the established
behaviour.

This fixees the bug of lineup and period occasionally adding to 31 with a red lineup clock.